update `Validator` type, use u63 type for power, update modules

+++ x/distribution/src/keeper/allocation.rs
@@ -146,7 +146,7 @@ impl<
             //
             // Ref: https://github.com/cosmos/cosmos-sdk/issues/2525#issuecomment-430838701
 
-            let power_fraction = Decimal256::from_atomics(vote.validator.power, 0)?
+            let power_fraction = Decimal256::from_atomics(u64::from(vote.validator.power), 0)?
                 .checked_div(Decimal256::from_atomics(total_previous_power, 0)?)?
                 .floor();
             let reward = fee_multiplier.checked_mul_dec_truncate(power_fraction)?;
